TypeSystemContext: drop the second variant of isMarkedNullable
Before this commit, we had two versions of this function, with
KotlinTypeMarker and SimpleTypeMarker receivers respectively.
However, most use-sites have RigidTypeMarker as an actual receiver type,
so the first version was used widely. As it delegated to the second one
with "this is SimpleTypeMarker" additional interface type check,
we suffered some performance problems here (this place is hot enough).
In fact, it's more effective to do such type checks in overrides,
as they may use a class type check instead and it's faster.

#KT-71353 Fixed